Fix scheduled migration and queries

+++ db/migrations/20240519231400_create_scheduled_tournament_table.sql
@@ -2,7 +2,7 @@
 CREATE TABLE scheduled_tournaments (
     id BIGSERIAL PRIMARY KEY,
     name TEXT NOT NULL,
-    entry_fee INT NOT NULL,
+    entry_fee BIGINT NOT NULL,
     start_time TIMESTAMP NOT NULL,
     recurrence_pattern TEXT NOT NULL,
     recurrence_start_timestamp TIMESTAMP NOT NULL,
